third_party/WebKit/LayoutTests/vr/latest/requestDevice_always_resolves.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>

promise_test(t => navigator.vr.requestDevice().then(device => {
  assert_true(device == null);
}), "Test that navigator.vr.requestDevice always resolves with at least a null.");

</script>
